major code cleanup
