+++ roofit/histfactory/CMakeLists.txt
@@ -37,6 +37,7 @@ ROOT_STANDARD_LIBRARY_PACKAGE(HistFactory
     RooStats/HistFactory/RooBarlowBeestonLL.h
     RooStats/HistFactory/Sample.h
     RooStats/HistFactory/Systematics.h
+    RooStats/HistFactory/HistFactoryJSONTool.h                    
   SOURCES
     src/Asimov.cxx
     src/Channel.cxx
@@ -62,7 +63,8 @@ ROOT_STANDARD_LIBRARY_PACKAGE(HistFactory
     src/RooBarlowBeestonLL.cxx
     src/Sample.cxx
     src/Systematics.cxx
-    src/RYMLFactories.cxx            
+    src/RYMLFactories.cxx
+    src/HistFactoryJSONTool.cxx                
   DICTIONARY_OPTIONS
     "-writeEmptyRootPCM"
   DEPENDENCIES
