+++ roofit/histfactory/src/Measurement.cxx
@@ -118,7 +118,7 @@ void RooStats::HistFactory::Measurement::AddPreprocessFunction( std::string name
 }
 
 /// Returns a list of defined preprocess function expressions
-std::vector<std::string> RooStats::HistFactory::Measurement::GetPreprocessFunctions()
+std::vector<std::string> RooStats::HistFactory::Measurement::GetPreprocessFunctions() const
 {
 
   
@@ -249,229 +249,6 @@ void RooStats::HistFactory::Measurement::PrintTree( std::ostream& stream )
 
   std::cout << "End Measurement: " << GetName() << std::endl;
 
-}
-
-
-#ifdef INCLUDE_RYML
-#include <ryml.hpp>
-#include <c4/yml/std/map.hpp>
-#include <c4/yml/std/string.hpp>
-#include "RooStats/HistFactory/RooJSONFactoryWSTool.h"
-
-namespace RooStats { namespace HistFactory {
-    void read(c4::yml::NodeRef const& n, PreprocessFunction *v)
-    {
-      std::string name;
-      std::string expression;
-      std::string dependents;
-      std::string command;   
-
-      n["name"]       >> name;      
-      n["expression"] >> expression;
-      n["dependents"] >> dependents;
-      n["command"]    >> command;   
-      
-      v->SetName      (name      );
-      v->SetExpression(expression);
-      v->SetDependents(dependents);
-      v->SetCommand   (command   );
-    }
-    
-    void write(c4::yml::NodeRef *n, PreprocessFunction const& v)
-    {
-      *n |= c4::yml::MAP;
-
-      auto ch = n->append_child();
-      ch["name"] << v.GetName();
-      ch["expression"] << v.GetExpression();
-      ch["dependents"] << v.GetDependents();
-      ch["command"] << v.GetCommand();
-    }
-  }
-}
-
-namespace c4 { namespace yml {
-    template<class T> void read(NodeRef const& n, std::vector<T> *v){
-      for(size_t i=0; i<n.num_children(); ++i){
-        std::string e;
-        n[i]>>e;
-        v->push_back(e);
-      }
-    }
-    
-    template<class T> void write(NodeRef *n, std::vector<T> const& v){
-      *n |= c4::yml::SEQ;
-      for(auto e:v){
-        n->append_child() << e;
-      }
-    }
-  }
-}
-
-
-template<> void RooStats::HistFactory::Measurement::Export(c4::yml::NodeRef& n) const {
-  RooJSONFactoryWSTool::clearcache();
-  for(const auto& ch:this->fChannels){
-    if(!ch.CheckHistograms()) throw std::runtime_error("unable to export histograms, please call CollectHistograms first");
-  }
-
-  auto parlist = n["variables"];
-  parlist |= c4::yml::MAP;
-
-  auto pdflist = n["pdfs"];
-  pdflist |= c4::yml::MAP;
-
-  // collect information
-  std::map<std::string,RooStats::HistFactory::Constraint::Type> constraints;
-  std::map<std::string,NormFactor > normfactors;
-  for(const auto& ch:fChannels){
-    for(const auto&s: ch.GetSamples()){
-      for(const auto& sys:s.GetOverallSysList()){
-        constraints[sys.GetName()] = RooStats::HistFactory::Constraint::Gaussian;
-      }
-      for(const auto& sys:s.GetHistoSysList()){
-        constraints[sys.GetName()] = RooStats::HistFactory::Constraint::Gaussian;
-      }
-      for(const auto& sys:s.GetShapeSysList()){
-        constraints[sys.GetName()] = sys.GetConstraintType();
-      }
-      for(const auto& norm:s.GetNormFactorList()){
-        normfactors[norm.GetName()] = norm;
-      }
-    }
-  }
-
-  // parameters
-  
-  auto lumi = parlist["Lumi"];
-  lumi |= c4::yml::MAP;  
-  lumi["value"] << fLumi;  
-  lumi["relErr"] << fLumiRelErr;
-
-  for(const auto& par:fPOI){
-    auto node = parlist[c4::to_csubstr(par)];
-    node |= c4::yml::MAP;    
-    if(fParamValues.find(par) != fParamValues.end()){
-      node["value"] << fParamValues.at(par);
-    }    
-  }
-  for(const auto& par:fParamValues){
-    auto node = parlist[c4::to_csubstr(par.first)];
-    node |= c4::yml::MAP;    
-    node["value"] << par.second;
-    if(std::find(fConstantParams.begin(),fConstantParams.end(),par.first) != fConstantParams.end()){
-      node["const"] << 1;
-    }
-  }
-  
-  for(const auto& norm:normfactors){
-    auto node = parlist[c4::to_csubstr(norm.second.GetName())];
-    node |= c4::yml::MAP;
-    node["value"] << norm.second.GetVal();        
-    node["min"] << norm.second.GetLow();
-    node["max"] << norm.second.GetHigh();
-    if(norm.second.GetConst()){
-      node["const"] << norm.second.GetConst();
-    }
-  }
-  
-  // pdfs
-  for(const auto& sys:constraints){
-    auto node = pdflist[c4::to_csubstr(sys.first)];
-    node |= c4::yml::MAP;
-    node["type"] << RooStats::HistFactory::Constraint::Name(sys.second);
-    if(sys.second == RooStats::HistFactory::Constraint::Gaussian){
-      std::string xname = std::string("alpha_")+sys.first;
-      auto xpar = parlist[c4::to_csubstr(RooJSONFactoryWSTool::incache(xname))];
-      xpar |= c4::yml::MAP;
-      xpar["value"] << 0;      
-      xpar["max"] << 5.;
-      xpar["min"] << -5.;
-      xpar["err"] << 1;            
-      node["x"] << xname;
-      node["mean"] << "0.";
-      node["sigma"] << "1.";
-    }
-  }
-
-  for(const auto& sys:fGammaSyst){
-    auto node = pdflist[c4::to_csubstr(sys.first)];
-    node |= c4::yml::MAP;
-    node["value"] << sys.second;    
-  }
-  for(auto sys:fUniformSyst){
-    auto node = pdflist[c4::to_csubstr(sys.first)];
-    node |= c4::yml::MAP;
-    node["value"] << sys.second;    
-  }
-  for(auto sys:fLogNormSyst){
-    auto node = pdflist[c4::to_csubstr(sys.first)];    
-    node |= c4::yml::MAP;
-    node["value"] << sys.second;    
-  }
-  for(auto sys:fNoSyst){
-    auto node = pdflist[c4::to_csubstr(sys.first)];    
-    node |= c4::yml::MAP;    
-    node["value"] << sys.second;
-  }
-
-  if(fFunctionObjects.size() > 0){
-    auto funclist = n["functions"];
-    funclist |= c4::yml::MAP;    
-    funclist << fFunctionObjects;
-  }
-
-  // and finally, the simpdf
-  
-  auto sim = pdflist[c4::to_csubstr(this->GetName())];
-  sim |= c4::yml::MAP;
-  sim["type"] << "simultaneous";
-  auto simdict = sim["dict"];
-  simdict |= c4::yml::MAP;  
-  simdict["InterpolationScheme"] << fInterpolationScheme;  
-  simdict["toplevel"] << true;
-  auto ch = sim["pdfs"];
-  ch |= c4::yml::MAP;  
-  auto chlist = sim["channels"];  
-  chlist |= c4::yml::MAP;
-  for(const auto& c:fChannels){
-    c.Export(ch);
-    auto chentry = chlist[c4::to_csubstr(RooJSONFactoryWSTool::incache(c.GetName()))];
-    chentry << c.GetName();
-  }
-}
-#endif
-
-void RooStats::HistFactory::Measurement::PrintJSON( std::ostream& os ) {
-#ifdef INCLUDE_RYML  
-  ryml::Tree t;
-  c4::yml::NodeRef n = t.rootref();
-  n |= c4::yml::MAP;
-  this->Export(n);
-  os << c4::yml::as_json(t);
-#else
-  std::cerr << "JSON export only support with rapidyaml!" << std::endl;
-#endif
-}
-void RooStats::HistFactory::Measurement::PrintJSON( std::string filename ) {
-  std::ofstream out(filename);
-  this->PrintJSON(out);
-}
-
-void RooStats::HistFactory::Measurement::PrintYAML( std::ostream& os ) {
-#ifdef INCLUDE_RYML  
-  ryml::Tree t;
-  c4::yml::NodeRef n = t.rootref();
-  n |= c4::yml::MAP;
-  this->Export(n);
-  os << t;
-#else
-  std::cerr << "YAML export only support with rapidyaml!" << std::endl;
-#endif
-}
-void RooStats::HistFactory::Measurement::PrintYAML( std::string filename ) {
-  std::ofstream out(filename);
-  this->PrintYAML(out);
 }
