+++ roofit/histfactory/src/Sample.cxx
@@ -309,74 +309,6 @@ void RooStats::HistFactory::Sample::PrintXML( std::ofstream& xml ) {
 }
 
 
-#ifdef INCLUDE_RYML
-#include <ryml.hpp>
-#include <c4/yml/std/map.hpp>
-#include <RooStats/HistFactory/RooJSONFactoryWSTool.h>
-#include <c4/yml/std/string.hpp>
-
-namespace RooStats { namespace HistFactory {
-    template<> void RooStats::HistFactory::Sample::Export(c4::yml::NodeRef& n) const {
-      std::vector<std::string> obsnames;
-      obsnames.push_back("obs_x_"+fChannelName);
-      obsnames.push_back("obs_y_"+fChannelName);
-      obsnames.push_back("obs_z_"+fChannelName);
-      
-      auto s = n[c4::to_csubstr(fName)];
-      s |= c4::yml::MAP;
-      s["type"] << "histogram";
-      
-      if(fOverallSysList.size() > 0){
-        auto overallSys = s["overallSystematics"];
-        overallSys |= c4::yml::MAP;
-        for(const auto& sys:fOverallSysList){
-          auto node = overallSys[c4::to_csubstr(sys.GetName())];
-          node |= c4::yml::MAP;        
-          node["parameter"] << std::string("alpha_")+sys.GetName();
-          node["low"] << sys.GetLow();
-          node["high"] << sys.GetHigh();
-        }
-      }
-
-      if(fNormFactorList.size()>0){
-        auto normFactors = s["normFactors"];
-        normFactors |= c4::yml::SEQ;
-        for(auto& sys:fNormFactorList){
-          normFactors.append_child() << sys.GetName();
-        }
-      }
-
-      if(fHistoSysList.size()>0){
-        auto histoSys = s["histogramSystematics"];
-        histoSys |= c4::yml::MAP;
-        for(size_t i=0; i<fHistoSysList.size(); ++i){
-          auto sys = fHistoSysList[i];
-          auto node = histoSys[c4::to_csubstr(sys.GetName())];
-          node |= c4::yml::MAP;        
-          node["parameter"] << std::string("alpha_")+sys.GetName();
-          auto dataLow = node["dataLow"];
-          auto dataHigh = node["dataHigh"];          
-          RooJSONFactoryWSTool::exportHistogram(*(sys.GetHistoLow()),dataLow,obsnames);
-          RooJSONFactoryWSTool::exportHistogram(*(sys.GetHistoHigh()),dataHigh,obsnames);                    
-        }
-      }
-
-      // std::vector< RooStats::HistFactory::ShapeSys >    fShapeSysList;
-      // std::vector< RooStats::HistFactory::ShapeFactor > fShapeFactorList;
-  
-      auto tags = s["dict"];
-      tags |= c4::yml::MAP;      
-      tags["normalizeByTheory"] << fNormalizeByTheory;
-      tags["statErrorActivate"] << fStatErrorActivate;
-
-      auto data = s["data"];
-      RooJSONFactoryWSTool::exportHistogram(*(fhNominal.GetObject()),data,obsnames);
-    }
-  }
-}
-#endif
-
-
 // Some helper functions
 // (Not strictly necessary because
 //  methods are publicly accessable)
